fix(): Filter duplicate semantic versions
The 3.45.1 release of SQLite has multiple tags for the same semantic version, an additional tag with a typo is present. This was causing it incorrectly report there was a newer version present. Add a filter that checks to see if the version is present. The `semver.Version.Equals` method does not take into account any metadata so just convert to a string and compare that.